reorganize lance
