+++ langroid/agent/special/__init__.py
@@ -1,3 +1,7 @@
+from .relevance_extractor_agent import (
+    RelevanceExtractorAgent,
+    RelevanceExtractorAgentConfig,
+)
 from .doc_chat_agent import DocChatAgent, DocChatAgentConfig
 from .retriever_agent import (
     RecordMetadata,
@@ -5,38 +9,42 @@ from .retriever_agent import (
     RetrieverAgentConfig,
     RetrieverAgent,
 )
+from .lance_doc_chat_agent import LanceDocChatAgent
 from .table_chat_agent import (
     dataframe_summary,
     TableChatAgent,
     TableChatAgentConfig,
     RunCodeTool,
 )
-from .relevance_extractor_agent import (
-    RelevanceExtractorAgent,
-    RelevanceExtractorAgentConfig,
-)
 from . import sql
-from . import lance_rag
+from . import relevance_extractor_agent
 from . import doc_chat_agent
 from . import retriever_agent
+from . import lance_tools
+from . import lance_doc_chat_agent
+from . import lance_rag
 from . import table_chat_agent
 
 __all__ = [
+    "RelevanceExtractorAgent",
+    "RelevanceExtractorAgentConfig",
     "DocChatAgent",
     "DocChatAgentConfig",
     "RecordMetadata",
     "RecordDoc",
     "RetrieverAgentConfig",
     "RetrieverAgent",
+    "LanceDocChatAgent",
     "dataframe_summary",
     "TableChatAgent",
     "TableChatAgentConfig",
     "RunCodeTool",
-    "RelevanceExtractorAgent",
-    "RelevanceExtractorAgentConfig",
     "sql",
-    "lance_rag",
+    "relevance_extractor_agent",
     "doc_chat_agent",
     "retriever_agent",
+    "lance_tools",
+    "lance_doc_chat_agent",
+    "lance_rag",
     "table_chat_agent",
 ]
