+++ langroid/agent/special/lance_rag/__init__.py
@@ -1,10 +1,8 @@
-from . import lance_tools
 from . import query_planner_agent
 from . import critic_agent
 from . import lance_rag_task
 
 __all__ = [
-    "lance_tools",
     "query_planner_agent",
     "critic_agent",
     "lance_rag_task",
